backend/requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
babel==2.17.0
bcrypt==3.2.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.3
dnspython==2.7.0
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.20.1
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.115.12
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
googlesearch-python==1.3.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.0.2
oauthlib==3.3.1
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
playsound==1.3.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyMuPDF==1.26.0
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.2
reportlab==4.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
xlrd==2.0.1
